Add build dependencies to INSTALL file
Summary:
While trying to build the applet I followed the instructions in the INSTALL file but I got lots of error while running cmake.
Lots of googling told me which packages to install, but I thought it would be useful to have the list before I started, so I've added them to the file.
The package names are based on KDE Neon (Ubuntu 16.04 LTS).

Test Plan:
On a fresh installation, install the list of packages in the INSTALL file.
Run the cmake command in the INSTALL file and ensure it completes without errors.

Reviewers: martinkostolny, mart

Reviewed By: mart

Subscribers: ngraham, aacid, plasma-devel

Tags: #documentation, #plasma

Differential Revision: https://phabricator.kde.org/D13418